updates curriculums link to full curriculum
