+++ src/components/curriculum/CurriculumGraphQLAPI.js
@@ -6,9 +6,6 @@ import { Col, Row } from '../layout/Grid'
 import GraphQLServerDayOneSessions from './sessions/GraphQLServerDayOneSessions'
 import NodejsSession from './sessions/NodejsSession'
 // import GraphQLServerDayTwoSessions from './sessions/GraphQLServerDayTwoSessions'
-
-import { LinkButton } from '../buttons'
-import SectionCTA from './SectionCTA'
 import { GRAPHQL_API } from '../../config/data'
 import selectCurriculumLayout, { LIST_TWO_COL } from './selectCurriculumLayout'
 import { curriculumCommonPropTypes } from './'
@@ -19,7 +16,7 @@ const CurriculumGraphQLAPI = ({
   enableToggle,
   isOpen,
   toggleNavigateTo = `/graphql/curriculum?tab=${GRAPHQL_API}`,
-  showLinkToCurriculum = false,
+  showLinkToCurriculum = true,
   marketingCard,
   trainings,
 }) => {
@@ -63,13 +60,6 @@ const CurriculumGraphQLAPI = ({
       >
         <GraphQLServerDayTwoSessions />
       </Section> */}
-      {showLinkToCurriculum ? (
-        <SectionCTA>
-          <LinkButton to={`/graphql/curriculum?tab=${GRAPHQL_API}`}>
-            Full curriculum
-          </LinkButton>
-        </SectionCTA>
-      ) : null}
       {marketingCard}
     </React.Fragment>
   )
@@ -96,6 +86,7 @@ const CurriculumGraphQLAPI = ({
         layout,
         type,
         trainings,
+        curriculumTo: showLinkToCurriculum ? toggleNavigateTo : undefined,
       })}
     </React.Fragment>
   )
