--- src/components/curriculum/SectionCTA.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-const SectionCTA = styled.div.attrs({
-  className: 'curriculum-cta',
-})`
-  padding-top: 50px;
-  padding-bottom: 25px;
-`
-export default SectionCTA
