+++ src/components/curriculum/selectCurriculumLayout.js
@@ -1,4 +1,7 @@
 import React from 'react'
+import styled from 'styled-components'
+
+import { LinkButton } from '../buttons'
 import { Col, Row } from '../layout/Grid'
 import { List } from './CurriculumSection'
 import { UpcomingTrainingSection } from '../training'
@@ -7,6 +10,13 @@ import CorpTrainingFacts from './CorpTrainingKeyFacts'
 export const LIST_LAYOUT = 'list'
 export const LIST_TWO_COL = 'listTwoCol'
 
+const SectionCTA = styled.div.attrs({
+  className: 'curriculum-cta',
+})`
+  padding-top: 50px;
+  padding-bottom: 25px;
+`
+
 const selectCurriculumLayout = ({
   firstHalf,
   secondHalf,
@@ -14,13 +24,21 @@ const selectCurriculumLayout = ({
   type = '',
   trainings,
   corpTrainingFacts,
+  curriculumTo,
 }) => {
+  const curriculumButtonSection = (
+    <SectionCTA>
+      <LinkButton to={curriculumTo}>Full curriculum</LinkButton>
+    </SectionCTA>
+  )
+
   if (layout === LIST_LAYOUT) {
     return (
       <React.Fragment>
         <List>
           {firstHalf}
           {secondHalf}
+          {curriculumTo && curriculumButtonSection}
         </List>
       </React.Fragment>
     )
@@ -31,6 +49,7 @@ const selectCurriculumLayout = ({
           <Row>
             <Col md={6} lg={5}>
               {firstHalf}
+              {curriculumTo && curriculumButtonSection}
             </Col>
             <Col md={6} lg={5} lgOffset={1}>
               {secondHalf}
@@ -46,6 +65,7 @@ const selectCurriculumLayout = ({
           <Col md={6} lg={5} lgOffset={1}>
             {firstHalf}
             {secondHalf}
+            {curriculumTo && curriculumButtonSection}
           </Col>
           <Col md={6} lg={5} lgOffset={1}>
             {trainings ? (
