+++ src/pages/graphql/curriculum.js
@@ -70,6 +70,13 @@ class GraphQLCurriculum extends React.Component {
   }
 
   render() {
+    const commonCurriculumProps = {
+      enableToggle: true,
+      showTitle: false,
+      layout: LIST_LAYOUT,
+      showLinkToCurriculum: false,
+    }
+
     return (
       <Layout>
         {({ trainings }) => {
@@ -313,9 +320,7 @@ class GraphQLCurriculum extends React.Component {
                               <Col lg={1} lgOffset={1} />
                               <Col lg={9}>
                                 <CurriculumGraphQLBootcamp
-                                  enableToggle={true}
-                                  showTitle={false}
-                                  layout={LIST_LAYOUT}
+                                  {...commonCurriculumProps}
                                   marketingCard={
                                     trainingBootcamp && (
                                       <MarketingCard
@@ -375,9 +380,7 @@ class GraphQLCurriculum extends React.Component {
                               <Col lg={1} lgOffset={1} />
                               <Col lg={9}>
                                 <CurriculumGraphQLAPI
-                                  enableToggle={true}
-                                  showTitle={false}
-                                  layout={LIST_LAYOUT}
+                                  {...commonCurriculumProps}
                                   marketingCard={
                                     trainingApi && (
                                       <MarketingCard
@@ -427,9 +430,7 @@ class GraphQLCurriculum extends React.Component {
                               <Col lg={1} lgOffset={1} />
                               <Col lg={9}>
                                 <CurriculumGraphQLApollo
-                                  enableToggle={true}
-                                  showTitle={false}
-                                  layout={LIST_LAYOUT}
+                                  {...commonCurriculumProps}
                                   marketingCard={
                                     trainingClient && (
                                       <MarketingCard
