+++ src/pages/react/curriculum.js
@@ -75,6 +75,13 @@ class ReactCurriculum extends React.Component {
   }
 
   render() {
+    const commonCurriculumProps = {
+      enableToggle: true,
+      showTitle: false,
+      layout: LIST_LAYOUT,
+      showLinkToCurriculum: false,
+    }
+
     return (
       <Layout>
         {({ trainings }) => {
@@ -324,10 +331,7 @@ class ReactCurriculum extends React.Component {
                               <Col lg={9}>
                                 {trainingBootcamp && (
                                   <CurriculumReactBootcamp
-                                    layout={LIST_LAYOUT}
-                                    enableToggle={true}
-                                    showTitle={false}
-                                    showLinkToCurriculum={false}
+                                    {...commonCurriculumProps}
                                     marketingCard={
                                       <MarketingCard
                                         heading="Next React Bootcamp"
@@ -375,10 +379,7 @@ class ReactCurriculum extends React.Component {
                               <Col lg={9}>
                                 {trainingFundamentals && (
                                   <CurriculumReactFundamentals
-                                    layout={LIST_LAYOUT}
-                                    enableToggle={true}
-                                    showTitle={false}
-                                    showLinkToCurriculum={false}
+                                    {...commonCurriculumProps}
                                     marketingCard={
                                       <MarketingCard
                                         heading="Next React Fundamentals"
@@ -425,10 +426,7 @@ class ReactCurriculum extends React.Component {
                               <Col lg={9}>
                                 {trainingAdvanced && (
                                   <CurriculumAdvancedReact
-                                    enableToggle={true}
-                                    showTitle={false}
-                                    layout={LIST_LAYOUT}
-                                    showLinkToCurriculum={false}
+                                    {...commonCurriculumProps}
                                     marketingCard={
                                       <MarketingCard
                                         heading="Next React Advanced Training"
@@ -474,9 +472,7 @@ class ReactCurriculum extends React.Component {
                               <Col lg={1} lgOffset={1} />
                               <Col lg={9}>
                                 <CurriculumPartTime
-                                  enableToggle={true}
-                                  showTitle={false}
-                                  layout={LIST_LAYOUT}
+                                  {...commonCurriculumProps}
                                   marketingCard={
                                     trainingPartTime && (
                                       <MarketingCard
